将codeConfig移动到index.js
